Move expand_value to yash_semantics::expansion
It should return errors in yash_semantics::expansion::Error so that
callers do not have to handle error kinds that do not actually happen.
